Improve test coverage: add conftest, store 404 test, bug docs
- Extract create_order fixture into conftest.py
- Add order_update_response schema for PATCH validation
- Add 404 test for nonexistent store order
- Document bugs found (schema type, missing f-string) in test_pet.py
- Revert unnecessary base_url change
- Clean up parametrize syntax
