+++ api_helpers.py
@@ -1,6 +1,6 @@
 import requests
 
-base_url = 'http://127.0.0.1:5000'
+base_url = 'http://localhost:5000'
 
 # GET requests
 def get_api_data(endpoint, params = {}):
